Bump the `protobuf` and `grpcio` versions
The `protobuf` version 4.25.x will go out of support soon(ish), on 31
Mar 2025. Since we don't need to keep compatibility to any legacy code,
we can bump it in advance as a preparation.

The `googleapis-common-protos` version was also bumped to 1.65.0, as the
previous version was too old to work with the new `protobuf` version.

https://protobuf.dev/support/version-support/#python

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>